refactor: Keep service only stuff in a standalone file
Otherwise we have problems during tests since Handlebars
template files are imported
