+++ src/ducks/budgetAlerts/index.js
@@ -1,18 +1,9 @@
 import logger from 'cozy-logger'
-import isEqual from 'lodash/isEqual'
 import { TRANSACTION_DOCTYPE, ACCOUNT_DOCTYPE, GROUP_DOCTYPE } from 'doctypes'
 import { getCategoryId } from 'ducks/categories/helpers'
 import sumBy from 'lodash/sumBy'
-import {
-  fetchCategoryAlerts,
-  updateCategoryAlerts
-} from 'ducks/settings/helpers'
-import CategoryBudgetNotificationView from './CategoryBudgetNotificationView'
-import { sendNotification } from 'cozy-notifications'
+import { fetchCategoryAlerts } from 'ducks/settings/helpers'
 import { startOfMonth, endOfMonth, addDays, format } from 'date-fns'
-
-const lang = process.env.COZY_LOCALE || 'en'
-const dictRequire = lang => require(`../../locales/${lang}`)
 
 const log = logger.namespace('category-alerts')
 
@@ -95,7 +86,7 @@ export const fetchExpensesForAlert = async (client, alert, currentDate) => {
  *
  * @return {CategoryBudgetAlert}  - Updated alert (lastNotificationDate, lastNotificationAmount)
  */
-const collectAlertInfo = async (client, alert, options) => {
+export const collectAlertInfo = async (client, alert, options) => {
   const expenses = await fetchExpensesForAlert(
     client,
     alert,
@@ -129,80 +120,4 @@ const collectAlertInfo = async (client, alert, options) => {
   }
 }
 
-/**
- * Collects notification data for all alerts
- *
- * Return nulls if nothing is to be sent
- */
-const buildNotificationData = async (client, alerts, options = {}) => {
-  if (alerts.length === 0) {
-    log('info', 'No category budget alerts, bailing out.')
-  }
-
-  const data = []
-  for (let alert of alerts) {
-    const info = { alert }
-    try {
-      const collectedInfo = await collectAlertInfo(client, alert, options)
-      if (collectedInfo) {
-        Object.assign(info, collectedInfo)
-      }
-    } catch (e) {
-      log(
-        'error',
-        `Error while checking budget alert ${alert.id} (message: ${e.message})`
-      )
-    }
-    data.push(info)
-  }
-
-  const updatedAlerts = data.map(x => x.alert)
-  if (isEqual(alerts, updatedAlerts) && !options.force) {
-    log('info', 'No change to alerts, no need to send')
-    return null
-  }
-
-  return data
-}
-
-const buildNotificationView = client => {
-  const notifView = new CategoryBudgetNotificationView({
-    client,
-    lang,
-    data: {},
-    locales: {
-      [lang]: dictRequire(lang)
-    }
-  })
-  return notifView
-}
-
-/**
- * Sends category budget notification and updates settings if successful
- */
-const runCategoryBudgetService = async (client, options) => {
-  log('info', 'Running category budget notification service')
-  const alerts = await fetchCategoryAlerts(client)
-  log('info', `Found ${alerts.length} alert(s)`)
-  const data = await buildNotificationData(client, alerts, {
-    force: options.force,
-    currentDate: options.currentDate
-  })
-  if (!data) {
-    log('info', 'Nothing to send, bailing out')
-    return
-  }
-  const notifView = buildNotificationView(client)
-  await sendNotification(client, notifView)
-  log('info', `Saving updated alerts`)
-  const updatedAlerts = data.map(x => x.alert)
-  log('info', 'Saving updated category alerts')
-  await updateCategoryAlerts(client, updatedAlerts)
-}
-
-export {
-  buildNotificationData,
-  buildNotificationView,
-  fetchCategoryAlerts,
-  runCategoryBudgetService
-}
+export { fetchCategoryAlerts }
